# docs/openapi.yaml
basePath: /api/v1
definitions:
...
  dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      remember_me:
        type: boolean
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      phone_required:
        type: boolean
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserDTO'
      verification_required:
        type: boolean
      verification_token:
        type: string
    type: object
...
  dto.RegisterRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      role:
        enum:
        - user
        type: string
      whatsapp_number:
        type: string
    required:
    - email
    - name
    - password
    - whatsapp_number
    type: object
...
  dto.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      password_confirmation:
        type: string
      token:
        type: string
    required:
    - new_password
    - password_confirmation
    - token
    type: object
  dto.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.SetVerificationPhoneRequest:
    properties:
      verification_token:
        type: string
      whatsapp_number:
        type: string
    required:
    - verification_token
    - whatsapp_number
    type: object
...
  dto.UpdatePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        minLength: 6
        type: string
    required:
    - current_password
    - new_password
    type: object
  dto.UpdateProfileRequest:
    properties:
      avatar:
        type: string
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      whatsapp_number:
        type: string
    type: object
...
  dto.UserDTO:
    properties:
      avatar:
        type: string
      badge_icon:
        type: string
      badge_name:
        type: string
      created_at:
        type: string
      email:
        type: string
      exp:
        type: integer
      gold_coins:
        type: integer
      id:
        type: integer
      is_premium:
        type: boolean
      level:
        type: integer
      name:
        type: string
      premium_until:
        type: string
      profile_theme:
        type: string
      role:
        type: string
      whatsapp_number:
        type: string
      whatsapp_verified:
        type: boolean
    type: object
...
  dto.VerifyPhoneRequest:
    properties:
      code:
        type: string
      verification_token:
        type: string
    required:
    - code
    - verification_token
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ruangtenang.id
    name: API Support
  description: API untuk aplikasi Ruang Tenang - Platform Kesehatan Mental
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ruang Tenang API
  version: "1.0"
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Request a password reset code via WhatsApp for the account identified
        by email
      parameters:
      - description: Forgot password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Request password reset
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Login user
      tags:
      - Auth
  /auth/me:
    get:
      description: Get authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Auth
  /auth/password:
    put:
      consumes:
      - application/json
      description: Update authenticated user's password
      parameters:
      - description: Update password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update password
      tags:
      - Auth
  /auth/profile:
    put:
      consumes:
      - application/json
      description: Update authenticated user's profile
      parameters:
      - description: Update profile request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Register new user
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using token
      parameters:
      - description: Reset password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Reset password
      tags:
      - Auth
  /auth/verification/phone:
    post:
      consumes:
      - application/json
      parameters:
      - description: Phone setup request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SetVerificationPhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Set WhatsApp number during phone verification
      tags:
      - Auth
  /auth/verification/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: Verification request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyPhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Verify WhatsApp OTP and finish login
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"